data_preprocessor_requirements.txt
pandas
mysql-connector-python
sqlalchemy
openpyxl
bs4
